Allow the use of the errno name to compare errno codes.
Update to allow a non-numeric character string for the code_should_be
value and use that string to obtain the errno value from the python
errno module.

This will allow tests to use the "name" of an errno instead of having
to determine the numeric value directly.

Ex:
 ${exdev}=   Evaluate    errno.EXDEV    modules=errno
 Link    ${FILE3}  ${LINK2}  code_should_be=${exdev}

can be replaced with

  Link    ${FILE3}  ${LINK2}  code_should_be=EXDEV