make: Check for python3
Check for python3 first, and if not found, fall back to the old name
python (which may be py3 or py2).

Fix the missing new line at the end of the generated Makefile.conf file.

Remove the implicit python in the makefile, since that is not known
until after configure.