make: Use a local python virtualenv for development targets
Add the 'make init' target to create a local virtualenv and install the
packages we need to lint, unit test, generate docs, and upload packages.

The VIRTUAL_ENV env var is used to pick the active venv (if one) when
running make, otherwise we use the local .venv created by 'make init'.