lint: Add code format check to lint testenv
Run the `black --check` command to ensure the code is properly formatted
when the lint check is run. The --check option is a dry-run; no files
are changed, but black will fail if a file is not properly formatted.

Run `tox -e format` to reform the files to fix any formatting issues.