Always log code, stdout, and stderr in run_program()
Currently the run_program() helper function logs only when the code is
non-zero, and even in that case, the exit code is not logged.

Change this to always log the exit code and stdout and stderr of
programs run to make it easier to troubleshoot.

Continue to log at the info level for now, and if this is too noisy in
the logs, later we can change these logs to debug level.

Add a set of pytests to check we are logging and returning the code,
stdout, and stderr when calling run_program().