Print log messages to stdout by default
The robot.api.logger is not supported when running under a remote
server, which is now our standard way to use the library.

The usual work-around for this situation is to print log messages to
stdout with print(), which are captured by the server and sent back in
the XML-RPC calls to the process running the robot tests.

Introduce a new utility module called "logger" to be used as our logging
API in the library. By default, send logging output to stdout using the
regular print() function.

Enable the robot.api.logger when the ROBOT_API_LOGGER environment
variable is "true" (or "yes", "1", "enabled").  This can be used to go
back to the old method when running locally. (Even when running locally,
stdout is captured, so log messages will not be lost if ROBOT_API_LOGGER
is not set.)