tests: Align logged fixture with robot.api.logger
I incorrectly assumed the robot.api.logger functions took a variable
number of message arguments, like print(), but in fact they take a
single positional argument for the log message.  This means, if we
change logging in the library, it can pass our unit tests, but info will
be missing in the logs when we actually run tests with robotframework.

Fix the logged fixture so it matches the function signatures in the
robot.api.logger.  While here, also capture the trace, warn, and error
log messages in case we need those in the future.